+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added "How to use" information to the "About" section
 - Discogs notes field is empty by default. It should not be automatically filled
 - Use bandcamp "publish_date" as a release date
+- Output generated "Submission notes" value
 
 ## Fixed